SystemUI: Restore the COMBINED_QS_HEADERS flag
Before QPR3, we utilized the COMBINED_QS_HEADERS SystemUI flag to control the visibility of the new headers, making them overlayable through changes introduced in https://gerrit.aospa.co/#/c/30218. With the QPR3 update, Google removed this flag, making the new 'Combined QS headers' the default without the possibility of deactivation. This posed a challenge for lower-powered devices, leading to performance issues and lag when accessing the QS panel. This commit restores the flag, allowing maintainers to disable the new headers and providing relief for weaker devices.

Co-authored-by: Richard Raya <rdxzv.dev@gmail.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>
Change-Id: I665dc9e80f9a1ad3a907b10462ddb71a5a0290d9